requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
openpyxl==3.1.2